api/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15
django-cors-headers>=4.4
django-filter>=24.3
python-dotenv>=1.0
psycopg2-binary>=2.9
pillow>=10.0
django-summernote>=0.8
